Updated PathFinding
Finds the location files, and goes through them until the path has been
complete.  Currently a little buggy, and on occasion will crash the
game, but that should be easy to solve.  Mob does not naturally spawn,
for testing purposes I added a spawning mechanic to the  Paradoximer.  I
will remove this when I get it all fixed up.  :)